apps/pipelines/pyproject.toml
[project]
name = "pipelines"
version = "0.1.0"
description = "Ticket ingestion and ETL pipelines"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.10.0",
    "httpx>=0.27.0",
    "ml-core",
    "pandas>=2.2.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "shared",
    "torch>=2.10.0",
    "tqdm>=4.65.0",
]

[tool.uv.sources]
ml-core = { workspace = true }
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
